RSpec: move tmpnam_spec.rb
Not entirely sure this spec should exist, but leaving it for now. We
should eventually figure out how to test it via an external interface,
if it matters.
